+++ jest.config.js
@@ -1,8 +1,3 @@
 module.exports = {
-  projects: [
-    {
-      displayName: 'test',
-      preset: '@jameslnewell/jest-preset-test'
-    }
-  ]
+  preset: '@jameslnewell/jest-preset-test',
 };