+++ src/components.tsx
@@ -1,5 +1,5 @@
 import styled from 'styled-components';
-import { ThemedValueOrValueMap } from './types';
+import {ThemedValueOrValueMap} from './types';
 import {
   ThemedMarginProps,
   ThemedPaddingProps,
@@ -18,14 +18,15 @@ import {
   pb,
   pl,
   marginProps,
-  paddingProps
+  paddingProps,
 } from './themed';
 
 interface InlineProps {
   inline?: boolean;
 }
 
-const inlineMixin = ({ inline }: InlineProps) => inline && 'display: inline-block;';
+const inlineMixin = ({inline}: InlineProps): string | undefined =>
+  (inline && 'display: inline-block;') || undefined;
 
 export type DeprecatedSpacingProps = InlineProps & {
   all?: ThemedValueOrValueMap;
@@ -40,13 +41,13 @@ export type DeprecatedSpacingProps = InlineProps & {
 /** @deprecated */
 export const Margin = styled.div<DeprecatedSpacingProps>`
   ${inlineMixin}
-  ${({ all }) => all && m(all)}
-  ${({ horizontal }) => horizontal && mx(horizontal)}
-  ${({ vertical }) => vertical && my(vertical)}
-  ${({ top }) => top && mt(top)}
-  ${({ right }) => right && mr(right)}
-  ${({ bottom }) => bottom && mb(bottom)}
-  ${({ left }) => left && ml(left)}
+  ${({all}) => all && m(all)}
+  ${({horizontal}) => horizontal && mx(horizontal)}
+  ${({vertical}) => vertical && my(vertical)}
+  ${({top}) => top && mt(top)}
+  ${({right}) => right && mr(right)}
+  ${({bottom}) => bottom && mb(bottom)}
+  ${({left}) => left && ml(left)}
 `;
 
 Margin.displayName = 'Margin';
@@ -54,13 +55,13 @@ Margin.displayName = 'Margin';
 /** @deprecated */
 export const Padding = styled.div<DeprecatedSpacingProps>`
   ${inlineMixin}
-  ${({ all }) => all && p(all)}
-  ${({ horizontal }) => horizontal && px(horizontal)}
-  ${({ vertical }) => vertical && py(vertical)}
-  ${({ top }) => top && pt(top)}
-  ${({ right }) => right && pr(right)}
-  ${({ bottom }) => bottom && pb(bottom)}
-  ${({ left }) => left && pl(left)}
+  ${({all}) => all && p(all)}
+  ${({horizontal}) => horizontal && px(horizontal)}
+  ${({vertical}) => vertical && py(vertical)}
+  ${({top}) => top && pt(top)}
+  ${({right}) => right && pr(right)}
+  ${({bottom}) => bottom && pb(bottom)}
+  ${({left}) => left && pl(left)}
 `;
 
 Padding.displayName = 'Padding';